refactor: 토큰 로직 수정
